requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.1.1
gunicorn==20.0.4
psycopg2-binary==2.9.9
python-decouple==3.4
openai
pytest
pytest-mock
djangorestframework